// apps/mobile/src/hooks/useTheme.d.ts
export declare const useTheme: () => {
    isDark: boolean;
    colors: {
        primary: string;
        primaryDark: string;
        secondary: string;
        background: string;
        surface: string;
        surfaceVariant: string;
        text: string;
        textSecondary: string;
        success: string;
        error: string;
        border: string;
    };
};
//# sourceMappingURL=useTheme.d.ts.map